Extracted instances batch endpoint to separate class
